Add serialiser for the LineItem instances
This class represents an individual item in the items array for the
checkout payload as defined by affirm here:

https://docs.affirm.com/Integrate_Affirm/Direct_API#initiate checkout

We define a seperate class here because we'll be rendering a collection
of items, instead of just one set of data.

We also use rails routes here from the spree engine. This means that
applications looking to use our extension will _have_ to set the host
parameter in `Spree::Core::Engine.routes.default_url_helpers` This
should be emphasized as it's usually inferred at runtime, even in
production.