Add the callback controller that confirms the payment
When the payment is successfully setup on Affirm, this is the
controller that handles the post to confirm that is send from Affirm.
It will create the payment for the order and move the order to the next
state to be confirmed by the user.

introduce CallbackHook::Base

This is the base class that will be the default CallbackHook that will
be used to create and update the payments and redirect to the correct
url. The hook can be defined by setting the
`SolidusAffirm::Config.callback_hook` to the class that inherits from
`CallbackHook::Base`