Introduce a helper for rendering the Affirm js config
Loading the configuration in the asset pipeline with dynamic erb parsing
code is not working well. Changing values in the api key for example
would mean that we have to restart the application. This commit
introduces a helper file and a partial that will render the correct
javascript setup needed for Affirm in the head. It assumes that there is
a `yied :head` in the application layout.