Introduce the Affirm Payment Gateway
This will add the payment gateway for Affirm to be added in the admin.
We need the Affirm public and private api key to communicate with
Affirm.

The affirm js url preference will be different in sandbox and production
environments, instead of hardcode these based on the Solidus environment
we provide a configuration option on the gateway to set it correctly.