Make the callback controller configurable
To provide maximum flexibility for using this extension, you can now
provide a controller to `SolidusAffirm::Config` that will be used for
handling the `post` to the `confirm` action and the `get` to the
`cancel` action.

To use your own controller·you can use the configuration like this:

`SolidusAffirm::Config.callback_controller_name = "my_controller"` and
then provide a controller in `Spree::MyController` that will have a
`confirm` and a `cancel` method.