Use the install generator to add the gateway to Solidus
To provide the users with the most flexible solution, this commit
removes the adding of the gateway to the array of gateways by default.
Instead it will use the install generator to add a line to the spree.rb
initializer that will add the Affirm gateway to the list of available
gateways to use.