Make Affirm checkouts non-reusable
This commit into Solidus core:
https://github.com/solidusio/solidus/commit/547d459c26b2f48bbe2a75b58c90063f678fced2

Introduced a check if payment sources are `reusable?`. The Checkout
model here is used as the source, but doesn't implement that method,
which causes any code that calls this to throw a method missing error.
In particular I've seen this occur on the /admin/payments/new page.