Fix checkout crashing when going back with Affirm
This fixes a crash that would occur when a user used the back button
(or one of the checkout step links) in the :confirm state and switched
their payment method to Affirm.

Since in this situation the order is already in the :confirm state
when Affirm redirects back to solidus, we are just skipping the
payment.order.next! transition that otherwise will raise an exception.