Use escaped unicode for single quotes
html data-attributes and javascript JSON parameters do not support
single quotes properly. This commit replaces all single quotes in the
Affirm data-attribute to the Unicoded escaped version. AMS already does this
out of the box for the & character. Making sure we do the same thing for single
quotes fixes the data and can properly create an Affirm payment now.

Fixes #42