Remove call to authorize payments in confirm
Payments are authorized when an order is completed as part of the
process! method in Spree::Payment::Processing. We want to wait until
then to authorize the payments to prevent any issues where an
amount is authorized and then the cart is abandoned or possible extra
authorizations.

Instead of authorizing we want to move the order to the confirm state
unless it is already in the confirm state. Previously this was setting
the payment amount to be the authorized amount but we are no longer
pre-authorizing. The payment amount will now be the full order amount.